patch capabilityCheck & AddWebhookHandler
